third_party/WebKit/LayoutTests/media/track/text-track-cue-exceptions.html

<!DOCTYPE html>
<title>Tests that TextTrackCue exceptions are properly messaged to developers.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
function getExceptionMsg(property) {
    return "Failed to set the '" + property
        + "' property on 'TextTrackCue': The provided double value is non-finite.";
}
var cue = new VTTCue(0, 0, "Test.");
function testProperty(property) {
    var expected_exceptioin_msg = getExceptionMsg(property);
    test(function() {
        var testValues = [Number.NaN, Number.POSITIVE_INFINITY, Number.NEGATIVE_INFINITY];
        for (var i = 0; i < testValues.length; i++) {
            assert_throws(new TypeError(), function() { cue[property] = testValues[i]; });
            try {
                cue[property] = testValues[i];
            } catch (e) {
                assert_equals(e.message, expected_exceptioin_msg);
            }
        }
    }, expected_exceptioin_msg);
}
testProperty("startTime");
testProperty("endTime");
</script>
